remove providers from components
