+++ packages/components/src/props.tsx
@@ -1,7 +1,15 @@
-import { ApolloClient, NormalizedCacheObject } from '@apollo/client'
+import { NextApiRequestCookies } from 'next/dist/server/api-utils'
+import Cookies from 'universal-cookie'
+import { setContext } from '@apollo/client/link/context'
+import {
+  createHttpLink,
+  InMemoryCache,
+  ApolloClient,
+  NormalizedCacheObject,
+} from '@apollo/client'
+import decode, { JwtPayload } from 'jwt-decode'
 import { GetServerSidePropsContext, GetServerSidePropsResult } from 'next'
-import { APOLLO_STATE_PROP_NAME, getClient } from './client'
-import { COOKIE_ADDRESS, COOKIE_JWT_TOKEN } from './session'
+import { COOKIE_JWT_TOKEN } from './session'
 
 type UserProp = {
   address: string | null
@@ -16,6 +24,50 @@ type GetServerSidePropsContextWithUser = GetServerSidePropsContext & {
   user: UserProp
 }
 
+export const APOLLO_STATE_PROP_NAME = '__APOLLO_STATE__'
+
+function getClient(
+  uri: string,
+  ssrMode: boolean,
+  cookies: NextApiRequestCookies | undefined,
+): ApolloClient<NormalizedCacheObject> {
+  const httpLink = createHttpLink({
+    uri,
+  })
+
+  // this function is called every time apollo is making a request
+  const authLink = setContext((_, context) => {
+    const c = new Cookies(cookies)
+
+    const jwtToken = c.get(COOKIE_JWT_TOKEN)
+    if (!jwtToken) return context
+
+    // check expiration of jwt token
+    const res = decode<JwtPayload>(jwtToken)
+    if (res.exp && res.exp < Math.ceil(Date.now() / 1000)) return context
+
+    return {
+      ...context,
+      headers: {
+        ...context.headers,
+        authorization: 'Bearer ' + jwtToken,
+      },
+    }
+  })
+
+  return new ApolloClient({
+    link: authLink.concat(httpLink),
+    cache: new InMemoryCache({
+      typePolicies: {
+        Account: {
+          keyFields: ['address'],
+        },
+      },
+    }),
+    ssrMode,
+  })
+}
+
 export function wrapServerSideProps<T extends { [key: string]: unknown }>(
   url: string,
   handler: (
@@ -27,15 +79,15 @@ export function wrapServerSideProps<T extends { [key: string]: unknown }>(
 ) => Promise<GetServerSidePropsResult<T>> {
   return async (context) => {
     const client = getClient(url, true, context.req.cookies)
+    const jwt = context.req.cookies[COOKIE_JWT_TOKEN]
+    const address = jwt
+      ? decode<JwtPayload & { address: string }>(jwt).address
+      : null
     const contextWithUser = {
       ...context,
       user: {
-        address: context.req.cookies[COOKIE_ADDRESS] || null,
-        token: context.req.cookies[COOKIE_ADDRESS]
-          ? context.req.cookies[
-              COOKIE_JWT_TOKEN(context.req.cookies[COOKIE_ADDRESS])
-            ] || null
-          : null,
+        address: address,
+        token: jwt || null,
       },
     }
     const result = await handler(contextWithUser, client)
